# Conflicts:
#	build.gradle
#	gradle.properties

+++ src/main/java/slimeknights/tconstruct/library/recipe/alloying/IMutableAlloyTank.java
@@ -13,5 +13,5 @@ public interface IMutableAlloyTank extends IAlloyTank {
   FluidStack drain(int tank, FluidStack fluidStack);
 
   /** Fills the tank with the output */
-  int fill(FluidStack fluidStack);
+  long fill(FluidStack fluidStack);
 }
